Added code to send setup to Konnected
Added code to talk to the Konnected device and provide it with the IO configuration for each pin based on the configuration from the file.

Also tweaked the app initialization to potentially work better when deploying, but this may change.

Code is currently disabled as startup is not the best place for this.  Particularly with debug mode and double-booting server, there are timing issues as the server and Konnected are restarting multiple times.

A connectivity heartbeat function should be created to test for liveness and this can potentially go there.